app/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
atomicwrites==1.4.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py==1.11.0
pydantic==2.12.5
pydantic_core==2.41.5
pytest==7.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.3.0
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
